app/requirements.txt
Django==4.1.7
Pillow==9.4.0
tzdata==2022.7
psycopg2-binary==2.9.5
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-pwa==1.1.0
gunicorn==20.1.0
django-request==1.6.2
django-ninja==0.20.0
crispy-bootstrap5==0.7
django-filter==23.1
